% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_labels_from_df.R
\name{apply_labels_from_df}
\alias{apply_labels_from_df}
\title{Apply labels to a data frame using another data frame}
\usage{
apply_labels_from_df(
  df,
  df_with_labels,
  col_name = "VariableName",
  col_label = "VariableLabel"
)
}
\arguments{
\item{df}{"Data frame to which labels will be applied"}

\item{df_with_labels}{"Data frame where labels are supplied"}

\item{col_name}{"Name of the column in df_with_labels where columns of df are supplied"}

\item{col_label}{"Name of the column in df_with_labels where the corresponding labels are supplied"}
}
\value{
A nice and pretty data frame with column labels
}
\description{
Apply labels to a data frame using another data frame
}
\examples{
tt <- apply_labels_from_df(world_seen_by_eu[, -12],
                           variables_world_seen_by_eu,
                           col_label = "VariableLabel_FR")
tt
labelled::var_label(tt)
}
