% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_name.R
\name{country_name}
\alias{country_name}
\title{Conversion entre nom complet de pays et leur code ISO2}
\usage{
country_name(list, from = c("name", "iso2"), lang = "fr")
}
\arguments{
\item{list}{Vecteur de classe caractère à convertir}

\item{from}{Point de départ de la conversion (name, ou iso2)}

\item{lang}{Langue du nom complet}
}
\description{
Conversion entre nom complet de pays et leur code ISO2
}
