% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib.R
\name{distrib}
\alias{distrib}
\title{Statistical Distribution Summary}
\usage{
distrib(var, precision = "summary", getLabel = T, useNA = "message")
}
\arguments{
\item{var}{Vector of numeric data.}

\item{precision}{A character specifying the level of precision for the statistics.
Options: "summary" or "S", "edges" or "E", "deciles" or "D", "deciles_edges" or "D+E". Default is "summary".}

\item{getLabel}{Logical argument - Should the labels of the variables be displayed in a message when printing the result of the function?}

\item{useNA}{When "message", shows the information about NAs in the variables in a message. When "include", shows it inside the returned data.frame.}
}
\value{
A data.frame object containing the selected descriptive statistics.

The function calculates the following statistics based on the precision option:
\itemize{
\item "summary": min, p25, p50 (median), p75, max, mean, sd
\item "edges": min, p01, p05, p10, p25, p50, p75, p90, p95, p99, max, mean, sd
\item "deciles": min, p10, p20, p30, p40, p50, p60, p70, p80, p90, max, mean, sd
\item "deciles_edges": min, p01, p05, p10, p20, p30, p40, p50, p60, p70, p80, p90, p95, p99, max, mean, sd
}

If useNA is set to "include", additional information about NAs is included in the returned data.frame.
\itemize{
\item "n_NA": Count of missing values
\item "p_NA": Proportion of missing values
}
}
\description{
This function computes various descriptive statistics for the given variable.
}
\examples{
# Example using the default precision ("summary")
data <- rnorm(100)
distrib(data)

# Example using precision "edges"
data <- rnorm(100)
distrib(data, precision = "edges")

}
