% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EUcountries.R
\docType{data}
\name{EUcountries}
\alias{EUcountries}
\title{EU countries and their labels}
\format{
\subsection{\code{EUcountries}}{

A data frame with 33 rows and 4 columns:
\describe{
\item{abreviation}{2 letter ISO country codes}
\item{country}{Nom du pays en anglais ou en langue origianle}
\item{FRcountry}{Nom du pays en français}
\item{ENcountry}{Nom du pays en anglais}
\item{EUmembership}{Pays-membre de l'Union européenne}
\item{currency}{Monnaie utilisée dans le paie (code ISO)}
...
}
}
}
\usage{
EUcountries
}
\description{
EU countries and their labels
}
\keyword{datasets}
