% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphDistrib.R
\name{graphDistrib}
\alias{graphDistrib}
\title{Graphical Distribution Summary}
\usage{
graphDistrib(var)
}
\arguments{
\item{var}{A vector of numeric data.}
}
\value{
A ggplot object representing the density plot.

The function also prints information about the variable, including the total number of observations,
the number of valid (non-NA) values, and the proportion of invalid (NA) values.
}
\description{
This function generates a density plot to visualize the distribution of a numeric variable.
}
\examples{
# Example using a numeric vector
data <- rnorm(100)
graphDistrib(data)

}
