% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphDistribCroisee.R
\name{graphDistribCroisee}
\alias{graphDistribCroisee}
\title{Graphique de distribution croisée}
\usage{
graphDistribCroisee(var1, var2)
}
\arguments{
\item{var1}{Vecteur de données numériques pour l'axe x.}

\item{var2}{Vecteur de données discrètes utilisé pour le croisement et la coloration.}
}
\value{
Un objet ggplot représentant le graphique de distribution croisée.
}
\description{
Cette fonction génère un graphique de distribution croisée pour une variable continue et une variable discrète.
}
\examples{
# Exemple d'utilisation de la fonction
data1 <- rnorm(100)
data2 <- sample(c("A", "B", "C"), 100, replace = TRUE)
graphDistribCroisee(data1, data2)

}
