% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_color_latex.R
\name{gt_color_latex}
\alias{gt_color_latex}
\title{Correction of rgba calls after as_latex}
\usage{
gt_color_latex(
  x,
  df_colors = tibble::tribble(~rgb, ~html, "rgba(0,255,0,0.2)", "#E0FFE0",
    "rgba(255,255,0,0.2)", "#FFFFE0", "rgba(0,0,139,0.2)", "#CCCCE8",
    "rgba(0,0,139,0.15)", "#D9D9EE", )
)
}
\arguments{
\item{df_colors}{data.frama with rgba calls and HTML color tag}

\item{A}{character vector of a gt LaTeX table}
}
\description{
Correction of rgba calls after as_latex
}
