% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_combine_vertical.R
\name{gt_combine_vertical}
\alias{gt_combine_vertical}
\title{Vertically combine gt tables within LaTeX environment}
\usage{
gt_combine_vertical(list_of_gt)
}
\arguments{
\item{list_of_gt}{A list of gt tables already converted with "as_latex"}
}
\description{
Vertically combine gt tables within LaTeX environment
}
