% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_fmt_latex.R
\name{gt_fmt_latex}
\alias{gt_fmt_latex}
\title{Formatting gt table(s) to LaTeX within a table environment}
\usage{
gt_fmt_latex(
  x,
  float = "h",
  tabcolsep = 0,
  arraystretch = 1,
  stretchtable = 1,
  normalstretch = 1.5
)
}
\arguments{
\item{x}{A (list of) gt table(s) already converted with "as_latex"}

\item{float}{Floating parameter for the table (defaults to "h")}

\item{tabcolsep}{Seperator between the colomns in pt (defaults to 0)}

\item{arraystretch}{Seperator between the rows in pt (defaults to 1)}

\item{stretchtable}{Setstretch parameter within the table}

\item{normalstretch}{Setstretch parameter outside the table}
}
\description{
Formatting gt table(s) to LaTeX within a table environment
}
