% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guillemetsFR.R
\name{guillemetsFR}
\alias{guillemetsFR}
\title{Guillemets français
type is "in" or "out" depending if it is opening or closing.}
\usage{
guillemetsFR(type = "none")
}
\description{
Guillemets français
type is "in" or "out" depending if it is opening or closing.
}
