% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkCreatoR.R
\name{networkCreatoR}
\alias{networkCreatoR}
\title{Interactive shiny interface to create and edit networks}
\usage{
networkCreatoR()
}
\description{
Interactive shiny interface to create and edit networks
}
