% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density2d_ellipses.R
\name{plot_density2d_ellipses}
\alias{plot_density2d_ellipses}
\title{Title}
\usage{
plot_density2d_ellipses(x, y, groups = NULL, ellipse_levels = 0.5, weights)
}
\arguments{
\item{weights}{}
}
\description{
Title
}
