% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_compa.R
\name{tbl_compa}
\alias{tbl_compa}
\title{Compare Proportions by Country}
\usage{
tbl_compa(
  x_vector,
  y_vector,
  w_vector = NULL,
  country_vector,
  countries,
  y_index = 1,
  useNA = "no",
  filter = NULL,
  fun_prop = questionr::rprop,
  n_funprop = FALSE,
  drop_funprop = FALSE,
  digits = 2,
  country_to = NULL,
  label_ensemble = "Ensemble"
)
}
\arguments{
\item{x_vector}{A vector of the main variable to be analyzed.}

\item{y_vector}{A vector of the secondary variable to be analyzed.}

\item{w_vector}{A vector of weights. If NULL, equal weights are assumed.}

\item{country_vector}{A vector indicating the country for each observation.}

\item{countries}{A character vector of country codes to be included in the analysis.}

\item{y_index}{The index of the y dimension to use in the output table. Default is 1.}

\item{useNA}{How to handle NA values. Options are 'no', 'ifany', or 'always'. Default is 'no'.}

\item{filter}{A logical vector indicating which observations to include. If NULL, all observations are included.}

\item{fun_prop}{A function to calculate proportions. Default is \code{questionr::rprop}.}

\item{digits}{Number of decimal places to include in the results. Default is 2.}

\item{country_to}{An optional target format for country codes.}

\item{label_ensemble}{Label for the combined set of countries. Default is "Ensemble".}
}
\value{
A data frame containing the proportion tables for each specified country and the combined set.
}
\description{
This function computes weighted tables of proportions for specified countries
and a combined set, optionally applying a filter to the data.
}
\examples{
# Example usage
x <- sample(c("Yes", "No"), 100, replace = TRUE)
y <- sample(c("A", "B"), 100, replace = TRUE)
w <- runif(100, 0.5, 1.5)
country <- sample(c("FR", "SE", "PT"), 100, replace = TRUE)
tbl_compa(x, y, w, country, countries = c("FR", "SE", "PT"))

}
