% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_condi_compa.R
\name{tbl_condi_compa}
\alias{tbl_condi_compa}
\title{Compare Proportions by Country using conditions}
\usage{
tbl_condi_compa(
  x_vector,
  w_vector = NULL,
  country_vector,
  countries,
  x_index = 1,
  useNA = "no",
  filter = NULL,
  fun_prop = questionr::rprop,
  drop_funprop = FALSE,
  country_to = NULL,
  label_ensemble = ifelse(identical(questionr::rprop, fun_prop), "Total", "Ensemble"),
  conditions = list(named = TRUE)
)
}
\arguments{
\item{x_vector}{A vector of the main variable to be analyzed.}

\item{w_vector}{A vector of weights. If NULL, equal weights are assumed.}

\item{country_vector}{A vector indicating the country for each observation.}

\item{countries}{A character vector of country codes to be included in the analysis.}

\item{x_index}{The index of the y dimension to use in the output table. Default is 1. 0 means you keep all}

\item{useNA}{How to handle NA values. Options are 'no', 'ifany', or 'always'. Default is 'always'.}

\item{filter}{A logical vector indicating which observations to include. If NULL, all observations are included.}

\item{fun_prop}{A function to calculate proportions. Default is \code{questionr::rprop}.}

\item{conditions}{A named list of logical vectors indicating the conditions for each subset.}
}
\description{
Compare Proportions by Country using conditions
}
\examples{
data("hdv2003", package = "questionr")
tbl_condi_compa(
  x_vector = hdv2003$hard.rock,
  w_vector = hdv2003$poids,
  country_vector = hdv2003$sexe,
  countries = c("Homme", "Femme"),
  x_index = 0,
  conditions = list("Fait du sport" = hdv2003$sport \%in\% "Oui")
  )
}
