% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_flagged.R
\name{var_flagged}
\alias{var_flagged}
\title{Combine a character or factor variable with its flags}
\usage{
var_flagged(x, x_flag, flag_meaning)
}
\arguments{
\item{x}{Main variable}

\item{x_flag}{Flag variable attached to main variable}

\item{flag_meaning}{Flag description under the form of named (numerical) vector (see examples)}
}
\value{
Factor vector
}
\description{
Combine a character or factor variable with its flags
}
\examples{
sample <- data.frame(variable_x = c("Male", "Female", "Female", NA, NA, "Male", "Female"),
                     variable_x_F = c(1, 1, 1, -1, -2, 1, 1))
flag_meaning <- c("Missing" = -1, "Not applicable" = -2)
var_flagged(x = sample$variable_x, x_flag = sample$variable_x_F,
            flag_meaning = c("Missing" = -1, "Not applicable" = -2))
}
