% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_world_seen_by_eu.R
\docType{data}
\name{variables_world_seen_by_eu}
\alias{variables_world_seen_by_eu}
\title{World seen the European Union - Variables}
\format{
\subsection{\code{variables_world_seen_by_eu}}{

A data frame with 12 rows and 3 columns:
\describe{
\item{VariableName}{Name of the variable}
\item{VariableLabel_EN}{Label of the variable in English}
\item{VariableLabel_FR}{Label of the variable in French}
...
}
}
}
\source{
Data version 2020 \url{https://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units/countries}
}
\usage{
variables_world_seen_by_eu
}
\description{
World seen the European Union - Variables
}
\keyword{datasets}
