% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_name.R
\name{country_name}
\alias{country_name}
\title{Conversion entre nom complet de pays et leur code ISO2}
\usage{
country_name(
  list,
  from = c("name_fr", "name_en", "name_de", "iso2", "iso3"),
  to = c("name_fr", "name_en", "name_de", "iso2", "iso3"),
  drop_levels = TRUE
)
}
\arguments{
\item{list}{Vecteur de classe caractère à convertir}

\item{from}{Point de départ de la conversion (name_fr, name_en, name_de, iso2, iso3)}

\item{to}{Point d'arrivée de la conversion (name_fr, name_en, name_de, iso2, iso3)}

\item{drop_levels}{Enlever les pays en trop ? par défaut, TRUE}
}
\description{
Conversion entre nom complet de pays et leur code ISO2
}
\examples{
country_name(c("FR", "PT"), from = "iso2", to = "name_en")


}
