% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribCroisee.R
\name{distribCroisee}
\alias{distribCroisee}
\title{Distribution croisée de statistiques descriptives}
\usage{
distribCroisee(var1, var2, large = FALSE)
}
\arguments{
\item{var1}{Vecteur de données numériques.}

\item{var2}{Vecteur de données utilisé pour le croisement.}

\item{large}{Un booléen indiquant si des statistiques supplémentaires doivent être calculées.}
}
\value{
Une table contenant des statistiques descriptives croisées.

Si \code{large} est TRUE, le tableau inclut les statistiques suivantes :
\itemize{
\item min, p01, p05, p20, p25, p40, p50, p60, p75, p80, p95, p99, max
\item mean : Moyenne
\item sd : Écart-type
\item n : Nombre total d'observations
\item NA(n) : Nombre d'observations manquantes
\item NA(\%) : Pourcentage d'observations manquantes
}

Si \code{large} est FALSE, le tableau inclut les statistiques suivantes :
\itemize{
\item min, p25, p50, p75, max
\item mean : Moyenne
\item sd : Écart-type
\item n : Nombre total d'observations
\item NA(n) : Nombre d'observations manquantes
}
}
\description{
Cette fonction calcule différentes statistiques descriptives croisées en fonction de l'argument \code{large}.
}
