% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphDistrib.R
\name{graphDistrib}
\alias{graphDistrib}
\title{Graphique de distribution}
\usage{
graphDistrib(var)
}
\arguments{
\item{var}{Vecteur de données numériques.}
}
\value{
Un objet ggplot représentant le graphique de distribution.
}
\description{
Cette fonction génère un graphique de distribution pour une variable donnée.
}
\examples{
# Exemple d'utilisation de la fonction
data <- rnorm(100)
graphDistrib(data)

}
