% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd.freq.R
\name{wtd.freq}
\alias{wtd.freq}
\title{Weighted frequency tables.}
\usage{
wtd.freq(x, weights, filter = TRUE, useNA = "ifany")
}
\arguments{
\item{x}{either a vector to be tabulated, or a table object}

\item{weights}{vector of weights, must be the same length as \code{x}}

\item{useNA}{in "always", "no", "ifany" (default)}
}
\value{
The result is an object of class data.frame.
}
\description{
Generate and format weighted frequency tables from a variable or a table, with percentages and formatting options.
}
\examples{
sample <- data.frame(
 variable_x = c("Male", "Female", "Female", NA, NA, "Male", "Female", "Male"),
 weights =         c(1,        1,        3,  1,  2,      4,        1,    NA))
wtd.freq(sample$variable_x, sample$weights)
}
